requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
celery==5.3.6
redis==5.0.4
kombu==5.3.4
django==4.2.11
psycopg2-binary==2.9.9
yara-python==4.5.0
pydantic==2.7.1
python-dotenv==1.0.1
asgiref==3.8.1
locust==2.28.0
